user_service/requirements.txt
fastapi[all]
uvicorn
sqlalchemy
psycopg2-binary
passlib
python-jose[cryptography]
python-multipart
alembic
asyncpg
pydantic-settings

pytest
pytest-asyncio
httpx
freezegun
pytest-cov
factory_boy